requirements.txt
pydantic>=1.9.0
tornado>=6.1
requests>=2.27.1
python-dateutil>=2.8.2